AddValidator transaction example
